refactor: use Buffer for partial write loop (more efficient)
Change from string slicing to Buffer-based offset writing.
This is more efficient as it avoids creating new string slices
on each iteration.

The buffer form of writeSync() takes an offset parameter that
specifies where in the buffer to start reading from:
  writeSync(fd, buffer, offset, length, position)

This is the standard Unix approach for handling partial writes.